drivers: clock_control: cleanup common part of renesas cpg driver
Move mstp and reg write rcar cpg API/arrays to an appropriate CPG
driver SoC file. Renesas CPG MSSR file contains only common driver
code and data.
