+++ drivers/clock_control/clock_control_renesas_cpg_mssr.h
@@ -39,9 +39,10 @@ struct renesas_cpg_mssr_data {
 
 	struct k_spinlock lock;
 
-	uint32_t (*get_div_helper)(uint32_t reg, uint32_t module);
-	uint32_t (*get_mul_helper)(uint32_t reg, uint32_t module);
-	int (*set_rate_helper)(uint32_t module, uint32_t *div, uint32_t *div_mask);
+	uint32_t (*get_div_helper)(mem_addr_t reg_addr, uint32_t module);
+	uint32_t (*get_mul_helper)(mem_addr_t reg_addr, uint32_t module);
+	int (*set_rate_helper)(const struct device *dev, struct cpg_clk_info_table *clk_info,
+			       uint32_t div);
 };
 
 #define RENESAS_CPG_NONE -1
@@ -75,46 +76,6 @@ struct renesas_cpg_mssr_data {
 		.children_list	= NULL,				\
 		.next_sibling	= NULL,				\
 	}
-
-#ifdef CONFIG_SOC_SERIES_RCAR_GEN3
-/* Software Reset Clearing Register offsets */
-#define SRSTCLR(i)      (0x940 + (i) * 4)
-
-/* CPG write protect offset */
-#define CPGWPR          0x900
-
-/* Realtime Module Stop Control Register offsets */
-static const uint16_t mstpcr[] = {
-	0x110, 0x114, 0x118, 0x11c,
-	0x120, 0x124, 0x128, 0x12c,
-	0x980, 0x984, 0x988, 0x98c,
-};
-
-/* Software Reset Register offsets */
-static const uint16_t srcr[] = {
-	0x0A0, 0x0A8, 0x0B0, 0x0B8,
-	0x0BC, 0x0C4, 0x1C8, 0x1CC,
-	0x920, 0x924, 0x928, 0x92C,
-};
-
-/* CAN-FD Clock Frequency Control Register */
-#define CANFDCKCR                 0x244
-
-/* Clock stop bit */
-#define CANFDCKCR_CKSTP           BIT(8)
-
-/* CANFD Clock */
-#define CANFDCKCR_PARENT_CLK_RATE 800000000
-#define CANFDCKCR_DIVIDER_MASK    0x1FF
-
-/* Peripherals Clocks */
-#define S3D4_CLK_RATE             66600000	/* SCIF	*/
-#define S0D12_CLK_RATE            66600000	/* PWM	*/
-#endif
-
-void rcar_cpg_write(uint32_t base_address, uint32_t reg, uint32_t val);
-
-int rcar_cpg_mstp_clock_endisable(uint32_t base_address, uint32_t module, bool enable);
 
 struct cpg_clk_info_table *renesas_cpg_find_clk_info_by_module_id(const struct device *dev,
 								  uint32_t domain,
